Extend min_value and max_value type size in pheno parquet reports
Some phenotype databases could not import as their values would exceed
float32

+++ dae/dae/pheno/pheno_import.py
@@ -1233,11 +1233,11 @@ def write_reports_to_parquet(
         ),
         pa.field(
             "min_value",
-            pa.float32(),
+            pa.float64(),
         ),
         pa.field(
             "max_value",
-            pa.float32(),
+            pa.float64(),
         ),
         pa.field(
             "values_domain",
